chore: move project db name template setting to settings tab
