+++ frontend/src/components/ProjectDeploymentConfigPanel.vue
@@ -14,61 +14,7 @@
       />
     </div>
 
-    <div class="text-lg font-medium leading-7 text-main">
-      {{ $t("project.db-name-template") }}
-    </div>
-    <div class="textinfolabel">
-      <i18n-t keypath="label.db-name-template-tips">
-        <template #placeholder>
-          <!-- prettier-ignore -->
-          <code v-pre class="text-xs font-mono bg-control-bg">{{DB_NAME}}__{{TENANT}}</code>
-        </template>
-        <template #link>
-          <a
-            class="normal-link inline-flex items-center"
-            href="https://www.bytebase.com/docs/change-database/batch-change/#specify-database-name-template"
-            target="__BLANK"
-          >
-            {{ $t("common.learn-more") }}
-            <heroicons-outline:external-link class="w-4 h-4 ml-1" />
-          </a>
-        </template>
-      </i18n-t>
-    </div>
-    <div class="mt-3 space-y-2">
-      <div>
-        <input
-          v-model="state.dbNameTemplate"
-          type="text"
-          class="textfield w-full"
-          :disabled="!state.isEditingDBNameTemplate"
-        />
-      </div>
-      <div class="flex items-center justify-end gap-x-2">
-        <button
-          v-if="!state.isEditingDBNameTemplate"
-          :disabled="!allowEdit"
-          class="btn-normal"
-          @click="beginEditDBNameTemplate"
-        >
-          {{ $t("common.edit") }}
-        </button>
-        <template v-if="state.isEditingDBNameTemplate">
-          <button class="btn-normal" @click="cancelEditDBNameTemplate">
-            {{ $t("common.cancel") }}
-          </button>
-          <button
-            class="btn-primary"
-            :disabled="!allowUpdateDBNameTemplate"
-            @click="confirmEditDBNameTemplate"
-          >
-            {{ $t("common.update") }}
-          </button>
-        </template>
-      </div>
-    </div>
-
-    <div class="text-lg font-medium leading-7 text-main mt-6 border-t pt-4">
+    <div class="text-lg font-medium leading-7 text-main mt-6 pt-4">
       {{ $t("common.deployment-config") }}
     </div>
 
@@ -149,7 +95,6 @@ import {
   pushNotification,
   useDeploymentConfigV1Store,
   useEnvironmentV1List,
-  useProjectV1Store,
 } from "@/store";
 import {
   DeploymentConfig,
@@ -169,8 +114,6 @@ type LocalState = {
   deployment: DeploymentConfig | undefined;
   originalDeployment: DeploymentConfig | undefined;
   error: string | undefined;
-  isEditingDBNameTemplate: boolean;
-  dbNameTemplate: string;
 };
 
 const props = defineProps({
@@ -197,8 +140,6 @@ const state = reactive<LocalState>({
   deployment: undefined,
   originalDeployment: undefined,
   error: undefined,
-  isEditingDBNameTemplate: false,
-  dbNameTemplate: props.project.dbNameTemplate,
 });
 
 const isDeploymentConfigDirty = computed((): boolean => {
@@ -310,36 +251,4 @@ watch(
   },
   { deep: true }
 );
-
-const allowUpdateDBNameTemplate = computed(() => {
-  return state.dbNameTemplate !== props.project.dbNameTemplate;
-});
-
-const beginEditDBNameTemplate = () => {
-  state.dbNameTemplate = props.project.dbNameTemplate;
-  state.isEditingDBNameTemplate = true;
-};
-
-const cancelEditDBNameTemplate = () => {
-  state.dbNameTemplate = props.project.dbNameTemplate;
-  state.isEditingDBNameTemplate = false;
-};
-
-const confirmEditDBNameTemplate = async () => {
-  try {
-    const projectPatch = cloneDeep(props.project);
-    projectPatch.dbNameTemplate = state.dbNameTemplate;
-    const updateMask = ["db_name_template"];
-    await useProjectV1Store().updateProject(projectPatch, updateMask);
-    pushNotification({
-      module: "bytebase",
-      style: "SUCCESS",
-      title: t("project.successfully-updated-db-name-template"),
-    });
-  } catch {
-    state.dbNameTemplate = props.project.dbNameTemplate;
-  } finally {
-    state.isEditingDBNameTemplate = false;
-  }
-};
 </script>
